back/requirements.txt
Flask
mysql-connector-python
flask-cors
PyJWT
boto3
requests
xmltodict
Pillow